# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Grondona API
  description: A Kotlin HTTP server for the PRODE Norteamérica 2026, built with Spring Boot.
  version: 1.0.0
  contact:
    name: Grondona API Support
    email: cristian.rana8@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: |
        Register a new user account. Returns a JWT token for immediate authentication.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              fullname: John Doe
              username: johndoe
              email: john@example.com
              password: secret123
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                userId: 550e8400-e29b-41d4-a716-446655440000
                username: johndoe
                email: john@example.com
                fullname: John Doe
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Username 'johndoe' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: username
                  rejectedValue: johndoe
    patch:
      tags:
        - Users
      summary: Update current user
      description: |
        Partially update the authenticated user's profile. 
        Only provided fields will be updated.
      operationId: updateUser
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              fullname: John Smith
              email: john.smith@example.com
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Email 'john@example.com' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: email
                  rejectedValue: john@example.com
  /api/users/login:
    post:
      tags:
        - Users
      summary: User login
      description: |
        Authenticate a user with user (username or email) and password. 
        Returns a JWT token for subsequent authenticated requests.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              user: johndoe
              password: secret123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 400
                error: Bad Request
                message: Invalid username or password
                timestamp: "2024-01-01T12:00:00"
  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: |
        Retrieve the profile of the currently authenticated user.
      operationId: getCurrentUser
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                fullname: John Doe
                username: johndoe
                email: john@example.com
                createdAt: "2024-01-01T00:00:00"
                updatedAt: "2024-01-01T00:00:00"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{userId}:
    delete:
      tags:
        - Users
      summary: Delete a user
      description: |
        Delete a user account. Users can only delete their own account.
      operationId: deleteUser
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID of the user to delete
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        "204":
          description: User deleted successfully
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cannot delete another user's account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 403
                error: Forbidden
                message: You can only delete your own account
                timestamp: "2024-01-01T12:00:00"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/me/groups:
    get:
      tags:
        - Users
      summary: Get my groups
      description: |
        Retrieve the list of groups the authenticated user has joined,
        each with its name and current member count.
      operationId: getMyGroups
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Groups retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserGroupResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tournaments:
    post:
      tags:
        - Tournaments
      summary: Create a new tournament
      description: |
        Create a new tournament by its name
      operationId: createTournament
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentRequest"
            example:
              name: Copa del Mundo 20206
      responses:
        "201":
          description: Tournament created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Tournament name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Tournament name 'Copa del Mundo 2026' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: name
                  rejectedValue: Copa del Mundo 20206
  /api/tournaments/{tournamentId}:
    get:
      tags:
        - Tournaments
      summary: Get a tournament by ID
      description: |
        Retrieve a specific tournament by its UUID.
      operationId: getTournament
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tournament retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentResponse"
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Tournaments
      summary: Update a tournament
      description: |
        Partially update a tournament. Only provided fields will be updated.
      operationId: updateTournament
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentRequest"
            example:
              name: Updated Tournament
              status: FINISHED
      responses:
        "200":
          description: Tournament updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Tournament name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: 409
                error: Conflict
                message: Tournament name 'Copa del Mundo 20206' already exists
                timestamp: "2024-01-01T12:00:00"
                data:
                  field: name
                  rejectedValue: Copa del Mundo 20206
    delete:
      tags:
        - Tournaments
      summary: Delete a tournament
      description: |
        Delete a tournament by its UUID.
      operationId: deleteTournament
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament to delete
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Tournament deleted successfully
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tournaments/{tournamentId}/matches:
    get:
      tags:
        - Matches
      summary: Get the list of matches for a tournament
      description: |
        Retrieve the list of matches for a tournament by its UUID.
      operationId: getTournamentMatches
      security:
        - bearerAuth: [ ]
      parameters:
        - name: tournamentId
          in: path
          required: true
          description: UUID of the tournament
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tournament matches retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentMatchesResponse"
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
        - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Invalid username or password
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T12:00:00"
        data:
          oneOf:
            - $ref: '#/components/schemas/ErrorConflictExtraData'
    ErrorConflictExtraData:
      type: object
      required:
        - field
        - rejected_value
      properties:
        field:
          type: string
          description: The field that caused the conflict
          example: username
        rejected_value:
          type: string
          description: The value that was rejected due to conflict
          example: johndoe
    CreateUserRequest:
      type: object
      required:
        - fullname
        - username
        - email
        - password
      properties:
        fullname:
          type: string
          description: Full name of the user
          minLength: 1
          example: John Doe
        username:
          type: string
          description: Unique username for authentication
          minLength: 3
          maxLength: 50
          example: johndoe
        email:
          type: string
          format: email
          description: Unique email address
          example: john@example.com
        password:
          type: string
          format: password
          description: User password (min 6 characters)
          minLength: 6
          example: secret123
    LoginRequest:
      type: object
      required:
        - user
        - password
      properties:
        user:
          type: string
          description: User for authentication (either username or email)
          example: johndoe
        password:
          type: string
          format: password
          description: User password
          example: secret123
    UpdateUserRequest:
      type: object
      description: All fields are optional. Only provided fields will be updated.
      properties:
        fullname:
          type: string
          description: New full name
          example: John Smith
        username:
          type: string
          description: New username (must be unique)
          minLength: 3
          maxLength: 50
          example: johnsmith
        email:
          type: string
          format: email
          description: New email address (must be unique)
          example: john.smith@example.com
        password:
          type: string
          format: password
          description: New password (min 6 characters)
          minLength: 6
          example: newsecret123
    AuthResponse:
      type: object
      required:
        - token
        - user_id
        - username
        - email
        - fullname
      properties:
        token:
          type: string
          description: JWT authentication token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user_id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        fullname:
          type: string
          description: Full name
          example: John Doe
    UserResponse:
      type: object
      required:
        - id
        - fullname
        - username
        - email
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        fullname:
          type: string
          description: Full name of the user
          example: John Doe
        username:
          type: string
          description: Username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
    UserGroupResponse:
      type: object
      description: A group the authenticated user has joined, with current member count
      required:
        - group_id
        - group_name
        - tournament_id
        - member_count
        - points
        - rank
        - role
      properties:
        group_id:
          type: string
          format: uuid
          description: Unique group identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        group_name:
          type: string
          description: Group name
          example: General
        tournament_id:
          type: string
          format: uuid
          description: Unique tournament identifier
          example: c0773be5-a5aa-4b88-9a3d-3143fae8d48f
        tournament_name:
          type: string
          description: Group name
          example: Copa del Mundo 2026
        member_count:
          type: integer
          format: int16
          description: Current number of members in the group
          example: 12
        points:
          type: number
          format: float16
          description: Amount of points scored in the group
          example: 43.7
        rank:
          type: integer
          format: int16
          description: Rank in the group
          example: 3
        role:
          type: string
          description: Role in the group
          example: MEMBER
...
    TournamentStatus:
      description: The status of a tournament
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - FINISHED
    CreateTournamentRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique tournament name
          maxLength: 100
          example: Copa del Mundo 2026
        status:
          $ref: '#/components/schemas/TournamentStatus'
    UpdateTournamentRequest:
      type: object
      description: All fields are optional. Only provided fields will be updated.
      properties:
        name:
          type: string
          description: New tournament name (must be unique)
          maxLength: 100
          example: Updated Tournament Name
        status:
          $ref: '#/components/schemas/TournamentStatus'
    TournamentResponse:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique tournament identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        name:
          type: string
          description: Unique tournament name
          maxLength: 100
          example: Copa del Mundo 20206
        status:
          $ref: '#/components/schemas/TournamentStatus'
    TeamResponse:
      type: object
      required:
        - id
        - name
        - code
        - status
        - icon
      properties:
        id:
          type: string
          format: uuid
          description: Unique team identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        name:
          type: string
          description: Unique team name
          maxLength: 100
          example: Argentina
        code:
          type: string
          description: Unique team code
          maxLength: 5
          example: ARG
        icon:
          type: string
          description: URL of the team icon
          example: https://flagicons.lipis.dev/flags/4x3/xx.svg
    MatchStatus:
      description: The status of a match
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - FINISHED
    TournamentMatchResponse:
      description: A match of a tournament
      type: object
      required:
        - id
        - match_key
        - home_team
        - away_team
        - home_quota
        - away_quota
        - tie_quota
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique match identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        match_key:
          type: string
          description: Unique identifier for the match within the tournament
          example: "QF2"
        home_team:
          $ref: '#/components/schemas/TeamResponse'
        away_team:
          $ref: '#/components/schemas/TeamResponse'
        home_quota:
          type: number
          format: float16
          description: The quota for a home team win
          example: 2.2
        away_quota:
          type: number
          format: float16
          description: The quota for an away team win
          example: 1.8
        tie_quota:
          type: number
          format: float16
          description: The quota for a tie
          example: 1.3
        status:
          $ref: '#/components/schemas/MatchStatus'
        substatus:
          type: string
        started_at:
          type: string
          format: date-time
          description: The timestamp for the match start
          example: '2026-06-12 13:30:00 +03:00'
        finished_at:
          type: string
          format: date-time
          description: The timestamp for the match end
          example: '2026-06-12 15:30:00 +03:00'
        home_goals:
          type: integer
          format: int8
          description: The amount of goals scored by the home team
          example: 1
        away_goals:
          type: integer
          format: int8
          description: The amount of goals scored by the away team
          example: 2
        home_penalties:
          type: integer
          format: int8
          description: The amount of penalties scored by the home team
          example: 6
        away_penalties:
          type: integer
          format: int8
          description: The amount of penalties scored by the away team
          example: 7
    TournamentMatchesResponse:
      description: All the matches of the tournament
      type: object
      required:
        - id
        - name
        - matches
      properties:
        id:
          type: string
          format: uuid
          description: Unique tournament identifier
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        name:
          type: string
          description: Tournament name
          example: Copa del Mundo 2026
        past_matches:
          type: array
          items:
            $ref: '#/components/schemas/TournamentMatchResponse'
        live_matches:
          type: array
          items:
            $ref: '#/components/schemas/TournamentMatchResponse'
        next_matches:
          type: array
          items:
            $ref: '#/components/schemas/TournamentMatchResponse'
...